feat: move query_balance from sdk to query
