+++ apps/extension/src/background/index.ts
@@ -7,7 +7,7 @@ import {
 import { defaultChainId, chains } from "@anoma/chains";
 import { init as initCrypto } from "@anoma/crypto/src/init";
 import { init as initShared } from "@anoma/shared/src/init";
-import { Sdk } from "@anoma/shared";
+import { Query, Sdk } from "@anoma/shared";
 
 import {
   ExtensionRouter,
@@ -61,7 +61,9 @@ const { REACT_APP_NAMADA_URL = DEFAULT_URL } = process.env;
 
   const sdkDataStr: string | undefined = await sdkStore.get(SDK_KEY);
 
+  //TODO: Most likely sdk and query should be a one thing
   const sdk = new Sdk(REACT_APP_NAMADA_URL);
+  const query = new Query(REACT_APP_NAMADA_URL);
   await sdk.fetch_masp_params();
 
   if (sdkDataStr) {
@@ -78,6 +80,7 @@ const { REACT_APP_NAMADA_URL = DEFAULT_URL } = process.env;
     extensionStore,
     defaultChainId,
     sdk,
+    query,
     cryptoMemory,
     requester
   );
